Remove creations
