--- models/all/src/creation.ts
@@ -1,31 +1,0 @@
-//
-// Copyright © 2020, 2021 Anticrm Platform Contributors.
-//
-// Licensed under the Eclipse Public License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License. You may
-// obtain a copy of the License at https://www.eclipse.org/legal/epl-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-
-import type { Client } from '@anticrm/core'
-import { createDeps as createTaskDeps } from '@anticrm/model-task'
-import { createDeps as createLeadDeps } from '@anticrm/model-lead'
-import { createDeps as createRecruitDeps } from '@anticrm/model-recruit'
-import { createDeps as createDemoDeps } from '@anticrm/model-demo'
-import { createDeps as createTrackerDeps } from '@anticrm/model-tracker'
-import { createDeps as createBoardDeps } from '@anticrm/model-board'
-
-export async function createDeps (client: Client): Promise<void> {
-  await createTaskDeps(client)
-  await createLeadDeps(client)
-  await createRecruitDeps(client)
-  await createDemoDeps(client)
-  await createTrackerDeps(client)
-  await createBoardDeps(client)
-}
