+++ models/recruit/src/index.ts
@@ -1,5 +1,5 @@
 //
-// Copyright © 2020, 2021 Anticrm Platform Contributors.
+// Copyright © 2022 Hardcore Engineering Inc.
 //
 // Licensed under the Eclipse Public License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License. You may
@@ -438,6 +438,5 @@ export function createModel (builder: Builder): void {
   createReviewModel(builder)
 }
 
-export { createDeps } from './creation'
 export { recruitOperation } from './migration'
 export { default } from './plugin'
