+++ models/recruit/src/migration.ts
@@ -1,5 +1,5 @@
 //
-// Copyright © 2020, 2021 Anticrm Platform Contributors.
+// Copyright © 2022 Hardcore Engineering Inc.
 //
 // Licensed under the Eclipse Public License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License. You may
@@ -13,55 +13,120 @@
 // limitations under the License.
 //
 
-import core, { Ref, TxOperations } from '@anticrm/core'
+import core, { Doc, Ref, Space, TxOperations } from '@anticrm/core'
 import { createOrUpdate, MigrateOperation, MigrationClient, MigrationUpgradeClient } from '@anticrm/model'
 import tags, { TagCategory } from '@anticrm/model-tags'
+import { createKanbanTemplate, createSequence } from '@anticrm/model-task'
 import { getCategories } from '@anticrm/skillset'
-import { createReviewTemplates, createSequence } from './creation'
+import { KanbanTemplate } from '@anticrm/task'
 import recruit from './plugin'
 
 export const recruitOperation: MigrateOperation = {
-  async migrate (client: MigrationClient): Promise<void> {
-  },
+  async migrate (client: MigrationClient): Promise<void> {},
   async upgrade (client: MigrationUpgradeClient): Promise<void> {
     const tx = new TxOperations(client, core.account.System)
+    await createDefaults(tx)
+  }
+}
 
-    await createSpace(tx)
+async function createDefaults (tx: TxOperations): Promise<void> {
+  await createSpace(tx)
 
+  await createOrUpdate(
+    tx,
+    tags.class.TagCategory,
+    tags.space.Tags,
+    {
+      icon: tags.icon.Tags,
+      label: 'Other',
+      targetClass: recruit.mixin.Candidate,
+      tags: [],
+      default: true
+    },
+    recruit.category.Other
+  )
+
+  for (const c of getCategories()) {
     await createOrUpdate(
       tx,
       tags.class.TagCategory,
       tags.space.Tags,
       {
         icon: tags.icon.Tags,
-        label: 'Other',
+        label: c.label,
         targetClass: recruit.mixin.Candidate,
-        tags: [],
-        default: true
+        tags: c.skills,
+        default: false
       },
-      recruit.category.Other
+      (recruit.category.Category + '.' + c.id) as Ref<TagCategory>
     )
-
-    for (const c of getCategories()) {
-      await createOrUpdate(
-        tx,
-        tags.class.TagCategory,
-        tags.space.Tags,
-        {
-          icon: tags.icon.Tags,
-          label: c.label,
-          targetClass: recruit.mixin.Candidate,
-          tags: c.skills,
-          default: false
-        },
-        (recruit.category.Category + '.' + c.id) as Ref<TagCategory>
-      )
-    }
-
-    await createReviewTemplates(tx)
-    await createSequence(tx, recruit.class.Review)
-    await createSequence(tx, recruit.class.Opinion)
   }
+
+  await createSequence(tx, recruit.class.Review)
+  await createSequence(tx, recruit.class.Opinion)
+  await createSequence(tx, recruit.class.Applicant)
+  await createDefaultKanbanTemplate(tx)
+  await createReviewTemplates(tx)
+}
+
+async function createReviewTemplates (tx: TxOperations): Promise<void> {
+  if ((await tx.findOne(core.class.TxCreateDoc, { objectId: recruit.template.Interview })) === undefined) {
+    await createKanbanTemplate(tx, {
+      kanbanId: recruit.template.Interview,
+      space: recruit.space.ReviewTemplates as Ref<Doc> as Ref<Space>,
+      title: 'Interview',
+      states: [
+        { color: 9, title: 'Prepare' },
+        { color: 10, title: 'Appointment' },
+        { color: 1, title: 'Opinions' }
+      ],
+      doneStates: [
+        { isWon: true, title: 'Pass' },
+        { isWon: false, title: 'Failed' }
+      ]
+    })
+  }
+
+  if ((await tx.findOne(core.class.TxCreateDoc, { objectId: recruit.template.Task })) === undefined) {
+    await createKanbanTemplate(tx, {
+      kanbanId: recruit.template.Task,
+      space: recruit.space.ReviewTemplates as Ref<Doc> as Ref<Space>,
+      title: 'Test task',
+      states: [
+        { color: 9, title: 'Prepare' },
+        { color: 10, title: 'Assigned' },
+        { color: 1, title: 'Review' },
+        { color: 4, title: 'Opinions' }
+      ],
+      doneStates: [
+        { isWon: true, title: 'Pass' },
+        { isWon: false, title: 'Failed' }
+      ]
+    })
+  }
+}
+
+async function createDefaultKanbanTemplate (tx: TxOperations): Promise<Ref<KanbanTemplate>> {
+  const defaultKanban = {
+    states: [
+      { color: 9, title: 'HR Interview' },
+      { color: 10, title: 'Technical Interview' },
+      { color: 1, title: 'Test task' },
+      { color: 0, title: 'Offer' }
+    ],
+    doneStates: [
+      { isWon: true, title: 'Won' },
+      { isWon: false, title: 'Lost' }
+    ]
+  }
+
+  return await createKanbanTemplate(tx, {
+    kanbanId: recruit.template.DefaultVacancy,
+    space: recruit.space.VacancyTemplates as Ref<Doc> as Ref<Space>,
+    title: 'Default vacancy',
+    states: defaultKanban.states,
+    doneStates: defaultKanban.doneStates
+  })
 }
 
 async function createSpace (tx: TxOperations): Promise<void> {
