+++ models/task/src/migration.ts
@@ -1,5 +1,5 @@
 //
-// Copyright © 2020, 2021 Anticrm Platform Contributors.
+// Copyright © 2022 Hardcore Engineering Inc.
 //
 // Licensed under the Eclipse Public License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License. You may
@@ -13,12 +13,88 @@
 // limitations under the License.
 //
 
-import { TxOperations } from '@anticrm/core'
-import {
-  MigrateOperation, MigrationClient, MigrationUpgradeClient
-} from '@anticrm/model'
+import { Class, Doc, Ref, Space, TxOperations } from '@anticrm/core'
+import { MigrateOperation, MigrationClient, MigrationUpgradeClient } from '@anticrm/model'
 import core from '@anticrm/model-core'
+import { KanbanTemplate, StateTemplate, DoneStateTemplate, genRanks, createKanban } from '@anticrm/task'
 import task from './plugin'
+
+/**
+ * @public
+ */
+export interface KanbanTemplateData {
+  kanbanId: Ref<KanbanTemplate>
+  space: Ref<Space>
+  title: KanbanTemplate['title']
+  states: Pick<StateTemplate, 'title' | 'color'>[]
+  doneStates: (Pick<DoneStateTemplate, 'title'> & { isWon: boolean })[]
+}
+
+/**
+ * @public
+ */
+export async function createSequence (tx: TxOperations, _class: Ref<Class<Doc>>): Promise<void> {
+  if ((await tx.findOne(task.class.Sequence, { attachedTo: _class })) === undefined) {
+    await tx.createDoc(task.class.Sequence, task.space.Sequence, {
+      attachedTo: _class,
+      sequence: 0
+    })
+  }
+}
+
+/**
+ * @public
+ */
+export async function createKanbanTemplate (
+  client: TxOperations,
+  data: KanbanTemplateData
+): Promise<Ref<KanbanTemplate>> {
+  const current = await client.findOne(task.class.KanbanTemplate, { _id: data.kanbanId })
+  if (current !== undefined) {
+    return current._id
+  }
+
+  const tmpl = await client.createDoc(
+    task.class.KanbanTemplate,
+    data.space,
+    {
+      doneStatesC: 0,
+      statesC: 0,
+      title: data.title
+    },
+    data.kanbanId
+  )
+
+  const doneStateRanks = [...genRanks(data.doneStates.length)]
+  await Promise.all(
+    data.doneStates.map((st, i) =>
+      client.addCollection(
+        st.isWon ? task.class.WonStateTemplate : task.class.LostStateTemplate,
+        data.space,
+        data.kanbanId,
+        task.class.KanbanTemplate,
+        'doneStatesC',
+        {
+          rank: doneStateRanks[i],
+          title: st.title
+        }
+      )
+    )
+  )
+
+  const stateRanks = [...genRanks(data.states.length)]
+  await Promise.all(
+    data.states.map((st, i) =>
+      client.addCollection(task.class.StateTemplate, data.space, data.kanbanId, task.class.KanbanTemplate, 'statesC', {
+        rank: stateRanks[i],
+        title: st.title,
+        color: st.color
+      })
+    )
+  )
+
+  return tmpl
+}
 
 async function createDefaultProject (tx: TxOperations): Promise<void> {
   const createTx = await tx.findOne(core.class.TxCreateDoc, {
@@ -60,12 +136,50 @@ async function createDefaultSequence (tx: TxOperations): Promise<void> {
   }
 }
 
+async function createDefaultKanbanTemplate (tx: TxOperations): Promise<Ref<KanbanTemplate>> {
+  const defaultKanban = {
+    states: [
+      { color: 9, title: 'Open' },
+      { color: 10, title: 'In Progress' },
+      { color: 1, title: 'Under review' },
+      { color: 0, title: 'Done' },
+      { color: 11, title: 'Invalid' }
+    ],
+    doneStates: [
+      { isWon: true, title: 'Won' },
+      { isWon: false, title: 'Lost' }
+    ]
+  }
+
+  return await createKanbanTemplate(tx, {
+    kanbanId: task.template.DefaultProject,
+    space: task.space.ProjectTemplates as Ref<Doc> as Ref<Space>,
+    title: 'Default project',
+    states: defaultKanban.states,
+    doneStates: defaultKanban.doneStates
+  })
+}
+
+async function createDefaultKanban (tx: TxOperations): Promise<void> {
+  const current = await tx.findOne(task.class.Kanban, {
+    attachedTo: task.space.TasksPublic
+  })
+  if (current !== undefined) return
+  const defaultTmpl = await createDefaultKanbanTemplate(tx)
+  await createKanban(tx, task.space.TasksPublic, defaultTmpl)
+}
+
+async function createDefaults (tx: TxOperations): Promise<void> {
+  await createDefaultSequence(tx)
+  await createDefaultProject(tx)
+  await createSequence(tx, task.class.Issue)
+  await createDefaultKanban(tx)
+}
+
 export const taskOperation: MigrateOperation = {
-  async migrate (client: MigrationClient): Promise<void> {
-  },
+  async migrate (client: MigrationClient): Promise<void> {},
   async upgrade (client: MigrationUpgradeClient): Promise<void> {
     const tx = new TxOperations(client, core.account.System)
-    await createDefaultSequence(tx)
-    await createDefaultProject(tx)
+    await createDefaults(tx)
   }
 }
